Implement end-to-end functional test
Also tidied up some exception handling code.  Function test methods are
not idempotent, so order of tests is mandated using
@FixMethodOrder(MethodSorters.NAME_ASCENDING) and adopting a naming
convention.  This is not ideal, but the tests will be run from maven,
which will start the application before running and stop it afterwards,
so the results will be predictable.